refactor[api-server]: lottery.link 패키지 위치 변경
