+++ src/main/java/com/watermelon/server/event/lottery/auth/interceptor/LoginCheckInterceptor.java
@@ -4,7 +4,7 @@ import com.watermelon.server.event.lottery.auth.exception.InvalidTokenException;
 import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
 import com.watermelon.server.event.lottery.auth.utils.AuthUtils;
 import com.watermelon.server.event.lottery.auth.exception.AuthenticationException;
-import com.watermelon.server.lottery.link.service.LinkService;
+import com.watermelon.server.event.lottery.link.service.LinkService;
 import com.watermelon.server.lottery.service.LotteryService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
