+++ src/test/java/com/watermelon/server/lottery/link/controller/LinkControllerTest.java
@@ -3,9 +3,10 @@ package com.watermelon.server.lottery.link.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.watermelon.server.ControllerTest;
 import com.watermelon.server.DocumentConstants;
-import com.watermelon.server.lottery.link.dto.MyLinkDto;
-import com.watermelon.server.lottery.link.service.LinkService;
-import com.watermelon.server.lottery.link.utils.LinkUtils;
+import com.watermelon.server.event.lottery.link.controller.LinkController;
+import com.watermelon.server.event.lottery.link.dto.MyLinkDto;
+import com.watermelon.server.event.lottery.link.service.LinkService;
+import com.watermelon.server.event.lottery.link.utils.LinkUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
